requirements.txt
attrs==25.1.0
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
Flask==3.1.0
Flask-Cors==5.0.0
Flask-SQLAlchemy==3.1.1
idna==3.10
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.5
lxml==5.3.1
MarkupSafe==3.0.2
platformdirs==4.3.6
psycopg2-binary==2.9.10
pytz==2025.1
requests==2.32.3
requests-file==2.1.0
requests-toolbelt==1.0.0
SQLAlchemy==2.0.38
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
zeep==4.3.1
